Move pieces out of the visual test, fix up comments, rewrite FinisherPiece.
