+++ osu.Game.Modes.Taiko/Objects/Drawable/Pieces/BashCirclePiece.cs
@@ -1,0 +1,33 @@
+using osu.Framework.Allocation;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Sprites;
+using osu.Framework.Graphics.Textures;
+using osu.Game.Graphics;
+
+namespace osu.Game.Modes.Taiko.Objects.Drawable.Pieces
+{
+    /// <summary>
+    /// A circle piece which is used to visualise Bash objects.
+    /// </summary>
+    public class BashCirclePiece : CirclePiece
+    {
+        private Sprite icon;
+
+        [BackgroundDependencyLoader]
+        private void load(OsuColour colours, TextureStore textures)
+        {
+            AccentColour = colours.YellowDark;
+
+            icon.Texture = textures.Get(@"Play/Taiko/bash-hit-inner");
+        }
+
+        protected override Framework.Graphics.Drawable CreateIcon()
+        {
+            return icon = new Sprite
+            {
+                Anchor = Anchor.Centre,
+                Origin = Anchor.Centre,
+            };
+        }
+    }
+}
