+++ osu.Game.Modes.Taiko/osu.Game.Modes.Taiko.csproj
@@ -50,6 +50,13 @@
     <Compile Include="Beatmaps\TaikoBeatmapConverter.cs" />
     <Compile Include="Beatmaps\TaikoBeatmapProcessor.cs" />
     <Compile Include="Judgements\TaikoJudgementInfo.cs" />
+    <Compile Include="Objects\Drawable\Pieces\BashCirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\CentreHitCirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\CirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\DrumRollCirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\FinisherPiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\RimHitCirclePiece.cs" />
+    <Compile Include="Objects\Drawable\Pieces\ScrollingCirclePiece.cs" />
     <Compile Include="TaikoDifficultyCalculator.cs" />
     <Compile Include="Objects\Drawable\DrawableTaikoHit.cs" />
     <Compile Include="Objects\TaikoBaseHit.cs" />
@@ -81,9 +88,7 @@
       <Name>osu.Game</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Objects\Drawable\Pieces\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
